swivel-api/requirements.txt
black==22.6.0
click==8.1.3
Flask==2.2.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
python-dotenv==0.20.0
tomli==2.0.1
typing_extensions==4.3.0
Werkzeug==2.2.1
zipp==3.8.1
